chore(shorebird_cli): rename command_runner.dart and command.dart to reflect the names of the classes they contain
