+++ packages/shorebird_cli/lib/src/commands/cache/cache_command.dart
@@ -1,5 +1,5 @@
-import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/commands/commands.dart';
+import 'package:shorebird_cli/src/shorebird_command.dart';
 
 /// {@template cache_command}
 /// `shorebird cache`
