+++ packages/shorebird_cli/lib/src/commands/cache/clean_cache_command.dart
@@ -3,9 +3,9 @@ import 'dart:io';
 
 import 'package:mason_logger/mason_logger.dart';
 import 'package:shorebird_cli/src/cache.dart';
-import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/platform.dart';
+import 'package:shorebird_cli/src/shorebird_command.dart';
 
 /// {@template clean_cache_command}
 /// `shorebird cache clean`
