+++ packages/shorebird_cli/lib/src/commands/flutter/versions/flutter_versions_command.dart
@@ -1,5 +1,5 @@
-import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/commands/commands.dart';
+import 'package:shorebird_cli/src/shorebird_command.dart';
 
 /// {@template flutter_versions_command}
 /// `shorebird flutter versions`
