+++ packages/shorebird_cli/lib/src/commands/upgrade_command.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:mason_logger/mason_logger.dart';
-import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/logger.dart';
+import 'package:shorebird_cli/src/shorebird_command.dart';
 import 'package:shorebird_cli/src/shorebird_version.dart';
 
 /// {@template upgrade_command}
